Renamed module to HomeView, configuration changes
