+++ src/modules/HomeView/components/EventList.js
