+++ src/modules/HomeView/components/Hero.js
