+++ src/modules/HomeView/config.json
