+++ src/modules/HomeView/redux/auth/reducer.js
