+++ src/modules/HomeView/redux/events/actions.js
