+++ src/modules/HomeView/redux/events/reducer.js
