+++ src/modules/HomeView/redux/hearings/reducer.js
