+++ src/modules/HomeView/sagas/hearings.js
@@ -2,14 +2,13 @@ import {
   put,
   call,
   takeLatest
-}                                                   from 'redux-saga/effects';
+} from 'redux-saga/effects';
 // import translations                                 from '../translations/general';
-
-import Config                                       from '../config.json';
-import makeRequest                                  from '../util/requests';
+import { getConfig } from '../config';
+import makeRequest from '../util/requests';
 import { default as HearingActions, HearingTypes }  from '../redux/hearings/actions';
 
-const HEARINGS_API_URL = Config.HEARINGS_API_BASE_URL
+const Config = getConfig();
 const LOCALE = 'fi'
 
 const getHeadlineForHearing = function(hearing) {
@@ -18,12 +17,12 @@ const getHeadlineForHearing = function(hearing) {
 
 const fetchHearings = function*() {
   try {
-    const response = yield call(makeRequest, HEARINGS_API_URL + "?open=true&ordering=-n_comments", 'GET', null)
+    const response = yield call(makeRequest, Config.HEARINGS_API_BASE_URL + "?open=true&ordering=-n_comments", 'GET', null)
     const hearings = response.results.map((hearing) => {
       return {
         key: hearing.id,
         imageUrl: hearing.main_image.url,
-        // headline: getHeadlineForHearing(hearing),
+        headline: getHeadlineForHearing(hearing),
         headline: 'hearings',
         urlSlug: hearing.slug
       }
