+++ src/modules/HomeView/sagas/index.js
@@ -1,0 +1,14 @@
+
+import { watchGetHeroContent, watchGetEvents, watchGetEvent } from './events';
+import { watchGetHearings } from './hearings';
+
+const sagas = function* sagas() {
+  yield [
+    watchGetHeroContent(),
+    watchGetHearings(),
+    watchGetEvents(),
+    watchGetEvent(),
+  ];
+};
+
+export default sagas;
