+++ src/modules/HomeView/util/popupNavigator.js
