+++ src/modules/HomeView/views/EventDetailView.js
