+++ src/modules/HomeView/views/HearingDetailView.js
